인증과 권한

+++ src/main/java/com/example/boardv1/board/BoardService.java
@@ -7,6 +7,7 @@ import org.springframework.transaction.annotation.Transactional;
 
 import lombok.RequiredArgsConstructor;
 
+// 책임 : 트랜잭션관리, DTO만들기, 권한체크(DB정보가 필요하니까)
 @RequiredArgsConstructor
 @Service
 public class BoardService {
@@ -17,18 +18,30 @@ public class BoardService {
         return boardRepository.findAll();
     }
 
-    public Board 상세보기(int id) {
+    public Board 수정폼게시글정보(int id, int sessionUserId) {
         Board board = boardRepository.findById(id)
                 .orElseThrow(() -> new RuntimeException("게시글을 찾을 수 없어요"));
 
+        if (sessionUserId != board.getUser().getId())
+            throw new RuntimeException("수정할 권한이 없습니다");
         return board;
     }
 
+    public BoardResponse.DetailDTO 상세보기(int id, Integer sessionUserId) {
+        Board board = boardRepository.findById(id)
+                .orElseThrow(() -> new RuntimeException("게시글을 찾을 수 없어요"));
+
+        return new BoardResponse.DetailDTO(board, sessionUserId);
+    }
+
     @Transactional // update, delete, insert 할때 붙이세요!!
-    public void 게시글수정(int id, String title, String content) {
-        //
+    public void 게시글수정(int id, String title, String content, int sessionUserId) {
         Board board = boardRepository.findById(id)
                 .orElseThrow(() -> new RuntimeException("수정할 게시글을 찾을 수 없어요"));
+
+        if (sessionUserId != board.getUser().getId())
+            throw new RuntimeException("수정할 권한이 없습니다");
+
         board.setTitle(title);
         board.setContent(content);
     }
@@ -51,10 +64,14 @@ public class BoardService {
     }
 
     @Transactional
-    public void 게시글삭제(int id) {
+    public void 게시글삭제(int id, int sessionUserId) {
         // 영속화
         Board board = boardRepository.findById(id)
                 .orElseThrow(() -> new RuntimeException("삭제할 게시글을 찾을 수 없어요"));
+
+        if (sessionUserId != board.getUser().getId())
+            throw new RuntimeException("삭제할 권한이 없습니다.");
+
         boardRepository.delete(board);
     } // 자동 flush
